git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@8078 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/CPUAdjustingSwapRequestInterval.java
@@ -27,8 +27,8 @@ public final class CPUAdjustingSwapRequestInterval implements SwapRequestInterva
         t.start();
     }
 
-    public synchronized double getValue() {
-        return currentValue;
+    public synchronized int getValue() {
+        return (int)currentValue;
     }
 
     public void run() {
