+++ src/freenet/node/SwapRequestInterval.java
@@ -4,5 +4,5 @@ package freenet.node;
  * Interface to get the current swap request interval
  */
 public interface SwapRequestInterval {
-    double getValue();
+    int getValue();
 }
